Remove outdated subscription event recording
and move back remaining methods from SubscriptionEventRecorder back to
PullRequestUpdater.
