+++ src/ProductConstructionService/ProductConstructionService.DependencyFlow/PullRequestUpdaters/CodeFlowPullRequestUpdater.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using Maestro.Common.Telemetry;
+using Maestro.Data;
 using Maestro.Data.Models;
 using Maestro.DataProviders;
 using Maestro.MergePolicies;
@@ -34,11 +35,11 @@ internal class CodeFlowPullRequestUpdater : PullRequestUpdater
     private readonly ICommentCollector _commentCollector;
     private readonly IPullRequestCommenter _pullRequestCommenter;
     private readonly IPullRequestStateManager _stateManager;
-    private readonly ISubscriptionEventRecorder _subscriptionEventRecorder;
     private readonly ICodeflowSourceDiffVerifier _codeflowSourceDiffVerifier;
     private readonly ISubscriptionUpdateOutcomeRecorder _outcomeRecorder;
     private readonly IPullRequestApprover _pullRequestApprover;
     private readonly IPullRequestTarget _target;
+    private readonly BuildAssetRegistryContext _context;
     private readonly ILogger<CodeFlowPullRequestUpdater> _logger;
 
     public CodeFlowPullRequestUpdater(
@@ -55,12 +56,12 @@ internal class CodeFlowPullRequestUpdater : PullRequestUpdater
         ICommentCollector commentCollector,
         IPullRequestCommenter pullRequestCommenter,
         IPullRequestStateManager stateManager,
-        ISubscriptionEventRecorder subscriptionEventRecorder,
         ICodeflowSourceDiffVerifier codeflowSourceDiffVerifier,
         ISubscriptionUpdateOutcomeRecorder outcomeRecorder,
         IPullRequestApprover pullRequestApprover,
+        BuildAssetRegistryContext context,
         ILogger<CodeFlowPullRequestUpdater> logger)
-        : base(target, mergePolicyEvaluator, remoteFactory, sqlClient, pullRequestCommenter, stateManager, subscriptionEventRecorder, outcomeRecorder, logger)
+        : base(target, mergePolicyEvaluator, remoteFactory, sqlClient, pullRequestCommenter, stateManager, outcomeRecorder, context, logger)
     {
         _vmrInfo = vmrInfo;
         _codeFlowExecutor = codeFlowExecutor;
@@ -74,10 +75,10 @@ internal class CodeFlowPullRequestUpdater : PullRequestUpdater
         _pullRequestCommenter = pullRequestCommenter;
         _logger = logger;
         _stateManager = stateManager;
-        _subscriptionEventRecorder = subscriptionEventRecorder;
         _codeflowSourceDiffVerifier = codeflowSourceDiffVerifier;
         _outcomeRecorder = outcomeRecorder;
         _pullRequestApprover = pullRequestApprover;
+        _context = context;
         _target = target;
     }
